誰も話していないときに4を返すように変更

+++ src/main/kotlin/providence/domain/Loudness.kt
@@ -1,3 +1,3 @@
 package providence.domain
 
-data class Loudness(val uid: Int, val time: Long, val loudness: Long)
+data class Loudness(val uid: Int, val time: Long, val loudness: Int)
